seperate model concerns
