--- sandbox/settings.py
@@ -1,116 +1,0 @@
-"""
-Django settings for td_biblio sandbox.
-"""
-
-import os
-
-import dj_database_url
-
-try:
-    from django.core.urlresolvers import reverse_lazy
-except:
-    from django.urls import reverse_lazy
-
-# Build paths inside the project like this: os.path.join(BASE_DIR, ...)
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-
-# Default secret key is for development only!
-SECRET_KEY = os.environ.get(
-    "DJANGO_SECRET_KEY", "2#jyyh#-tgk$%wdx0zgfx=-)#hr9ni(_m66a3!rmco8^a5cns!"
-)
-
-# SECURITY WARNING: don't run with debug turned on in production!
-DEBUG = os.environ.get("IS_PUBLIC_SANDBOX") is None
-
-ALLOWED_HOSTS = ["localhost", "127.0.0.1", "tailordev-biblio.herokuapp.com"]
-
-LOGIN_URL = reverse_lazy("login")
-LOGIN_REDIRECT_URL = reverse_lazy("td_biblio:import")
-
-# Application definition
-INSTALLED_APPS = [
-    "django.contrib.admin",
-    "django.contrib.auth",
-    "django.contrib.contenttypes",
-    "django.contrib.sessions",
-    "django.contrib.messages",
-    "whitenoise.runserver_nostatic",
-    "django.contrib.staticfiles",
-    "td_biblio",
-]
-
-MIDDLEWARE = (
-    "django.middleware.security.SecurityMiddleware",
-    "whitenoise.middleware.WhiteNoiseMiddleware",
-    "django.contrib.sessions.middleware.SessionMiddleware",
-    "django.middleware.common.CommonMiddleware",
-    "django.middleware.csrf.CsrfViewMiddleware",
-    "django.contrib.auth.middleware.AuthenticationMiddleware",
-    "django.contrib.messages.middleware.MessageMiddleware",
-    "django.middleware.clickjacking.XFrameOptionsMiddleware",
-)
-
-# Django < 1.10 compat
-MIDDLEWARE_CLASSES = (
-    "django.contrib.sessions.middleware.SessionMiddleware",
-    "django.middleware.common.CommonMiddleware",
-    "django.middleware.csrf.CsrfViewMiddleware",
-    "django.contrib.auth.middleware.AuthenticationMiddleware",
-    "django.contrib.auth.middleware.SessionAuthenticationMiddleware",
-    "django.contrib.messages.middleware.MessageMiddleware",
-    "django.middleware.clickjacking.XFrameOptionsMiddleware",
-)
-
-ROOT_URLCONF = "sandbox.urls"
-
-TEMPLATES = [
-    {
-        "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": [os.path.join(BASE_DIR, "templates")],
-        "APP_DIRS": True,
-        "OPTIONS": {
-            "context_processors": [
-                "django.template.context_processors.debug",
-                "django.template.context_processors.request",
-                "django.contrib.auth.context_processors.auth",
-                "django.contrib.messages.context_processors.messages",
-            ]
-        },
-    }
-]
-
-# Django < 1.8 compat
-TEMPLATE_DIRS = [os.path.join(BASE_DIR, "templates")]
-
-WSGI_APPLICATION = "sandbox.wsgi.application"
-
-
-# Database
-DATABASES = {"default": dj_database_url.config(conn_max_age=600)}
-
-
-# Password validation
-AUTH_PASSWORD_VALIDATORS = [
-    {
-        "NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator"  # noqa
-    },
-    {"NAME": "django.contrib.auth.password_validation.MinimumLengthValidator"},  # noqa
-    {"NAME": "django.contrib.auth.password_validation.CommonPasswordValidator"},  # noqa
-    {
-        "NAME": "django.contrib.auth.password_validation.NumericPasswordValidator"  # noqa
-    },
-]
-
-
-# Internationalization
-LANGUAGE_CODE = "en-us"
-TIME_ZONE = "UTC"
-USE_I18N = True
-USE_L10N = True
-USE_TZ = True
-
-
-# Static files (CSS, JavaScript, Images)
-STATIC_URL = "/static/"
-STATIC_ROOT = os.path.join(BASE_DIR, "public", "static")
-STATICFILES_DIRS = (os.path.join(BASE_DIR, "static"),)
