+++ td_biblio/models/author.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+from td_biblio.models.abstract_human import AbstractHuman
+
+
+class Author(AbstractHuman):
+    """Entry author"""
+
+    class Meta:
+        ordering = ("last_name", "first_name")
+        verbose_name = _("Author")
+        verbose_name_plural = _("Authors")
