+++ td_biblio/models/entry.py
@@ -1,132 +1,4 @@
 # -*- coding: utf-8 -*-
-from django.conf import settings
-from django.contrib.auth import get_user_model
-from django.db import models
-from django.utils.translation import ugettext_lazy as _
-
-
-class AbstractHuman(models.Model):
-    """Simple Abstract Human model
-
-    Note that this model may be linked to django registered users
-    """
-
-    first_name = models.CharField(_("First name"), max_length=100)
-    last_name = models.CharField(_("Last name"), max_length=100)
-    first_initial = models.CharField(_("First Initial(s)"), max_length=10, blank=True)
-
-    # This is a django user
-    user = models.ForeignKey(
-        settings.AUTH_USER_MODEL, blank=True, null=True, on_delete=models.CASCADE
-    )
-
-    alias = models.ForeignKey(
-        "self",
-        on_delete=models.CASCADE,
-        related_name="aliases",
-        related_query_name="alias_human",
-        blank=True,
-        null=True,
-    )
-
-    class Meta:
-        abstract = True
-
-    def __str__(self):
-        return self.get_formatted_name()
-
-    def save(self, *args, **kwargs):
-        """Set initials and try to set django user before saving"""
-
-        self._set_first_initial()
-        self._set_user()
-        super(AbstractHuman, self).save(*args, **kwargs)
-
-    def _set_first_initial(self, force=False):
-        """Set author first name initial"""
-
-        if self.first_initial and not force:
-            return
-        self.first_initial = " ".join([c[0] for c in self.first_name.split()])
-
-    def get_formatted_name(self):
-        """Return author formated full name, e.g. Maupetit J"""
-
-        return "%s %s" % (self.last_name, self.first_initial)
-
-    def _set_user(self):
-        """Look for local django user based on human name"""
-
-        if "" in (self.last_name, self.first_name):
-            return
-
-        self._set_first_initial()
-
-        User = get_user_model()
-        try:
-            self.user = User.objects.get(
-                models.Q(last_name__iexact=self.last_name),
-                models.Q(first_name__iexact=self.first_name)
-                | models.Q(first_name__istartswith=self.first_initial[0]),
-            )
-        except User.DoesNotExist:
-            pass
-        except User.MultipleObjectsReturned:
-            pass
-
-
-class Author(AbstractHuman):
-    """Entry author"""
-
-    class Meta:
-        ordering = ("last_name", "first_name")
-        verbose_name = _("Author")
-        verbose_name_plural = _("Authors")
-
-
-class Editor(AbstractHuman):
-    """Journal or book editor"""
-
-    class Meta:
-        ordering = ("last_name", "first_name")
-        verbose_name = _("Editor")
-        verbose_name_plural = _("Editors")
-
-
-class AbstractEntity(models.Model):
-    """Simple abstract entity"""
-
-    name = models.CharField(_("Name"), max_length=150)
-    abbreviation = models.CharField(
-        _("Entity abbreviation"),
-        max_length=100,
-        blank=True,
-        help_text=_("e.g. Proc Natl Acad Sci U S A"),
-    )
-
-    class Meta:
-        abstract = True
-
-    def __str__(self):
-        return self.name
-
-
-class Journal(AbstractEntity):
-    """Peer reviewed journal"""
-
-    class Meta:
-        verbose_name = _("Journal")
-        verbose_name_plural = _("Journals")
-
-
-class Publisher(AbstractEntity):
-    """Journal or book publisher"""
-
-    class Meta:
-        verbose_name = _("Publisher")
-        verbose_name_plural = _("Publishers")
-
-
 class Entry(models.Model):
     """The core model for references
 
@@ -377,40 +249,3 @@ class Entry(models.Model):
         queryset is not (M2M with a through case).
         """
         return [aer.author for aer in self.authorentryrank_set.all()]
-
-
-class Collection(models.Model):
-    """Define a collection of entries"""
-
-    name = models.CharField(_("Name"), max_length=100)
-    short_description = models.TextField(_("Short description"), blank=True, null=True)
-    entries = models.ManyToManyField("Entry", related_name="collections")
-
-    class Meta:
-        verbose_name = _("Collection")
-        verbose_name_plural = _("Collections")
-
-    def __str__(self):
-        return self.name
-
-
-class AuthorEntryRank(models.Model):
-    """Give the author rank for an entry author sequence"""
-
-    author = models.ForeignKey(Author, on_delete=models.CASCADE)
-    entry = models.ForeignKey(Entry, on_delete=models.CASCADE)
-    rank = models.IntegerField(
-        _("Rank"), help_text=_("Author rank in entry authors sequence")
-    )
-
-    class Meta:
-        verbose_name = _("Author Entry Rank")
-        verbose_name_plural = _("Author Entry Ranks")
-        ordering = ("rank",)
-
-    def __str__(self):
-        return "%(author)s:%(rank)d:%(entry)s" % {
-            "author": self.author,
-            "entry": self.entry,
-            "rank": self.rank,
-        }
